add compressed unsupervised learning data to archive root
The data files in this tutorial are large. Rather than use Git large file storage, which references the files in external storage, this approach keeps the individual data files within GitHub's file size limit. In addition, to keep the repository size down, the dataset is moved to the archive root and README references are updated to use the correct path.